converter_app_requirements.txt
streamlit
python-dotenv
gspread
oauth2client
langchain
langchain-google-genai
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
python-docx
gitpython
requests